+++ client/sdk/src/funcs/authCallback.ts
@@ -91,6 +91,7 @@ async function $do(
 
   const query = encodeFormQuery({
     "code": payload.code,
+    "state": payload.state,
   });
 
   const headers = new Headers(compactMap({