+++ client/sdk/src/funcs/authLogin.ts
@@ -3,8 +3,10 @@
  */
 
 import { GramCore } from "../core.js";
+import { encodeFormQuery } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
 import { compactMap } from "../lib/primitives.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { pathToFunc } from "../lib/url.js";
 import { GramError } from "../models/errors/gramerror.js";
@@ -30,6 +32,7 @@ import { Result } from "../types/fp.js";
  */
 export function authLogin(
   client: GramCore,
+  request?: operations.AuthLoginRequest | undefined,
   options?: RequestOptions,
 ): APIPromise<
   Result<
@@ -47,12 +50,14 @@ export function authLogin(
 > {
   return new APIPromise($do(
     client,
+    request,
     options,
   ));
 }
 
 async function $do(
   client: GramCore,
+  request?: operations.AuthLoginRequest | undefined,
   options?: RequestOptions,
 ): Promise<
   [
@@ -71,7 +76,23 @@ async function $do(
     APICall,
   ]
 > {
+  const parsed = safeParse(
+    request,
+    (value) =>
+      operations.AuthLoginRequest$outboundSchema.optional().parse(value),
+    "Input validation failed",
+  );
+  if (!parsed.ok) {
+    return [parsed, { status: "invalid" }];
+  }
+  const payload = parsed.value;
+  const body = null;
+
   const path = pathToFunc("/rpc/auth.login")();
+
+  const query = encodeFormQuery({
+    "redirect": payload?.redirect,
+  });
 
   const headers = new Headers(compactMap({
     Accept: "application/json",
@@ -97,6 +118,8 @@ async function $do(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
+    query: query,
+    body: body,
     userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);