+++ client/sdk/src/models/operations/serveimage.ts
@@ -8,24 +8,11 @@ import { safeParse } from "../../lib/schemas.js";
 import { Result as SafeParseResult } from "../../types/fp.js";
 import { SDKValidationError } from "../errors/sdkvalidationerror.js";
 
-export type ServeImageSecurity = {
-  apikeyHeaderGramKey?: string | undefined;
-  sessionHeaderGramSession?: string | undefined;
-};
-
 export type ServeImageRequest = {
   /**
    * The ID of the asset to serve
    */
   id: string;
-  /**
-   * Session header
-   */
-  gramSession?: string | undefined;
-  /**
-   * API Key header
-   */
-  gramKey?: string | undefined;
 };
 
 export type ServeImageResponse = {
@@ -34,93 +21,17 @@ export type ServeImageResponse = {
 };
 
 /** @internal */
-export const ServeImageSecurity$inboundSchema: z.ZodType<
-  ServeImageSecurity,
-  z.ZodTypeDef,
-  unknown
-> = z.object({
-  "apikey_header_Gram-Key": z.string().optional(),
-  "session_header_Gram-Session": z.string().optional(),
-}).transform((v) => {
-  return remap$(v, {
-    "apikey_header_Gram-Key": "apikeyHeaderGramKey",
-    "session_header_Gram-Session": "sessionHeaderGramSession",
-  });
-});
-
-/** @internal */
-export type ServeImageSecurity$Outbound = {
-  "apikey_header_Gram-Key"?: string | undefined;
-  "session_header_Gram-Session"?: string | undefined;
-};
-
-/** @internal */
-export const ServeImageSecurity$outboundSchema: z.ZodType<
-  ServeImageSecurity$Outbound,
-  z.ZodTypeDef,
-  ServeImageSecurity
-> = z.object({
-  apikeyHeaderGramKey: z.string().optional(),
-  sessionHeaderGramSession: z.string().optional(),
-}).transform((v) => {
-  return remap$(v, {
-    apikeyHeaderGramKey: "apikey_header_Gram-Key",
-    sessionHeaderGramSession: "session_header_Gram-Session",
-  });
-});
-
-/**
- * @internal
- * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
- */
-export namespace ServeImageSecurity$ {
-  /** @deprecated use `ServeImageSecurity$inboundSchema` instead. */
-  export const inboundSchema = ServeImageSecurity$inboundSchema;
-  /** @deprecated use `ServeImageSecurity$outboundSchema` instead. */
-  export const outboundSchema = ServeImageSecurity$outboundSchema;
-  /** @deprecated use `ServeImageSecurity$Outbound` instead. */
-  export type Outbound = ServeImageSecurity$Outbound;
-}
-
-export function serveImageSecurityToJSON(
-  serveImageSecurity: ServeImageSecurity,
-): string {
-  return JSON.stringify(
-    ServeImageSecurity$outboundSchema.parse(serveImageSecurity),
-  );
-}
-
-export function serveImageSecurityFromJSON(
-  jsonString: string,
-): SafeParseResult<ServeImageSecurity, SDKValidationError> {
-  return safeParse(
-    jsonString,
-    (x) => ServeImageSecurity$inboundSchema.parse(JSON.parse(x)),
-    `Failed to parse 'ServeImageSecurity' from JSON`,
-  );
-}
-
-/** @internal */
 export const ServeImageRequest$inboundSchema: z.ZodType<
   ServeImageRequest,
   z.ZodTypeDef,
   unknown
 > = z.object({
   id: z.string(),
-  "Gram-Session": z.string().optional(),
-  "Gram-Key": z.string().optional(),
-}).transform((v) => {
-  return remap$(v, {
-    "Gram-Session": "gramSession",
-    "Gram-Key": "gramKey",
-  });
 });
 
 /** @internal */
 export type ServeImageRequest$Outbound = {
   id: string;
-  "Gram-Session"?: string | undefined;
-  "Gram-Key"?: string | undefined;
 };
 
 /** @internal */
@@ -130,13 +41,6 @@ export const ServeImageRequest$outboundSchema: z.ZodType<
   ServeImageRequest
 > = z.object({
   id: z.string(),
-  gramSession: z.string().optional(),
-  gramKey: z.string().optional(),
-}).transform((v) => {
-  return remap$(v, {
-    gramSession: "Gram-Session",
-    gramKey: "Gram-Key",
-  });
 });
 
 /**